docs: 상품 위치 변경 (#269)
