+++ src/main/java/project/trendpick_pro/domain/recommend/repository/RecommendRepositoryCustom.java
@@ -2,7 +2,7 @@ package project.trendpick_pro.domain.recommend.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import project.trendpick_pro.domain.product.entity.dto.response.ProductListResponse;
+import project.trendpick_pro.domain.product.entity.product.dto.response.ProductListResponse;
 
 public interface RecommendRepositoryCustom {
     Page<ProductListResponse> findAllByMemberName(String username, Pageable pageable);
