+++ src/main/java/project/trendpick_pro/domain/review/entity/Review.java
@@ -1,19 +1,15 @@
 package project.trendpick_pro.domain.review.entity;
 
 import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.web.multipart.MultipartFile;
 import project.trendpick_pro.domain.common.base.BaseTimeEntity;
 import project.trendpick_pro.domain.common.file.CommonFile;
 import project.trendpick_pro.domain.member.entity.Member;
-import project.trendpick_pro.domain.product.entity.Product;
+import project.trendpick_pro.domain.product.entity.product.Product;
 import project.trendpick_pro.domain.review.entity.dto.request.ReviewSaveRequest;
-
-import java.util.List;
 
 @Entity
 @NoArgsConstructor
