+++ src/main/java/project/trendpick_pro/global/jmeter/JmeterController.java
@@ -11,7 +11,7 @@ import org.springframework.web.multipart.MultipartFile;
 import project.trendpick_pro.domain.common.base.rq.Rq;
 import project.trendpick_pro.domain.member.entity.Member;
 import project.trendpick_pro.domain.member.entity.dto.MemberInfoDto;
-import project.trendpick_pro.domain.product.entity.dto.request.ProductSaveRequest;
+import project.trendpick_pro.domain.product.entity.product.dto.request.ProductSaveRequest;
 import project.trendpick_pro.domain.product.service.ProductService;
 
 import java.io.IOException;
@@ -55,13 +55,13 @@ public class JmeterController {
 //        return orderService.showOrderItems(member, result.getData()).getData();
 //    }
 
-    @PostMapping("/edit")
-    public void modifyProduct(@RequestParam("productId") Long productId,
-                              @RequestParam("mainFile") MultipartFile mainFile,
-                              @RequestParam("subFile") List<MultipartFile> subFiles) throws IOException {
-        log.info("productId : {}", productId);
-        ProductSaveRequest productSaveRequest = new ProductSaveRequest("제목", "내용", "상의", "반소매티셔츠", "나이키", 50, 1000, List.of("오버핏청바지", "시티보이룩"));
-        Long id =  productService.modify(productId, productSaveRequest, mainFile, subFiles).getData();
-        log.info("id : {}", id);
-    }
+//    @PostMapping("/edit")
+//    public void modifyProduct(@RequestParam("productId") Long productId,
+//                              @RequestParam("mainFile") MultipartFile mainFile,
+//                              @RequestParam("subFile") List<MultipartFile> subFiles) throws IOException {
+//        log.info("productId : {}", productId);
+//        ProductSaveRequest productSaveRequest = new ProductSaveRequest("제목", "내용", "상의", "반소매티셔츠", "나이키", 50, 1000, List.of("오버핏청바지", "시티보이룩"));
+//        Long id =  productService.modify(productId, productSaveRequest, mainFile, subFiles).getData();
+//        log.info("id : {}", id);
+//    }
 }
